Refactor scripts to a console_script
Letting the Python package installer automatically
generate a console script is the official PyPA
recommendation for how to publish an entry point. (see
<https://packaging.python.org/guides/distributing-packages-using-setuptools/#scripts>)

This change adds support for invoking the ctx shell
command from an outside a virtual environment when the
shellctx package is installed inside a virtual
environment that is not currently activated. The
motivating use case for this change is
[pipx](https://pypa.github.io/pipx/), which is an
officially supported package of PyPA.